Modernize JavaScript build toolchain
Replace Rollup with Rolldown v1 for bundling (10-30x faster builds).
The new config lives in rolldown.config.mjs and the build script is
updated accordingly.

Replace ESLint + neostandard with Oxlint v1 for linting (50-100x faster).
The new config lives in .oxlintrc.json and is run via `pnpm lint`.
Use `args: none` on no-unused-vars to skip unused function parameters,
which serve as API documentation for callback signatures.
Rename .github/workflows/eslint.yml to lint.yml accordingly.

These choices are informed by ongoing upstream discussions about how
best to align the neostandard and Rollup ecosystems with modern
JavaScript tooling (neostandard/neostandard#350, rollup/plugins#2010).

Drop Babel — transpilation is now handled by Rolldown's built-in Oxc
transforms. The minimum browser targets are unchanged (Chrome 60+,
Firefox 60+, iOS 12+, Safari 12+).

Remove 9 unused npm devDependencies (`@babel/core`, `@babel/preset-env`,
`@rollup/plugin-babel`, `@rollup/plugin-node-resolve`, `eslint`,
`eslint-plugin-compat`, `neostandard`, `rollup`, `rollup-plugin-copy`).

Clean up .gitignore (remove stale !.babelrc negation) and update
Rakefile lint task to call the new `pnpm lint` script instead of
`pnpm eslint`.
