+++ .github/workflows/lint.yml
@@ -1,4 +1,4 @@
-name: ESLint
+name: Lint
 
 on:
   push:
@@ -10,8 +10,8 @@ permissions:
   contents: read
 
 jobs:
-  eslint:
-    name: ESLint
+  lint:
+    name: Lint
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v7
@@ -23,4 +23,4 @@ jobs:
     - name: Install node dependencies
       run: pnpm install
     - name: Run JavaScript linter
-      run: pnpm eslint
+      run: pnpm lint
