+++ .gitignore
@@ -17,6 +17,4 @@ test/log
 
 node_modules
 
-!.babelrc
-
 pnpm-lock.yaml
