+++ package.json
@@ -18,21 +18,15 @@
   },
   "homepage": "https://github.com/DavyJonesLocker/client_side_validations",
   "scripts": {
-    "build": "rollup -c",
+    "build": "rolldown -c",
+    "lint": "oxlint src/",
     "test": "test/javascript/run-qunit.mjs"
   },
   "devDependencies": {
-    "@babel/core": "^7.29.7",
-    "@babel/preset-env": "^7.29.7",
-    "@rollup/plugin-babel": "^7.0.0",
-    "@rollup/plugin-node-resolve": "^16.0.3",
     "chrome-launcher": "^1.2.1",
-    "eslint": "^9.39.4",
-    "eslint-plugin-compat": "^7.0.2",
-    "neostandard": "^0.13.0",
-    "puppeteer-core": "^25.1.0",
-    "rollup": "^4.60.4",
-    "rollup-plugin-copy": "^3.5.0"
+    "oxlint": "^1.72.0",
+    "puppeteer-core": "^25.3.0",
+    "rolldown": "^1.1.4"
   },
   "main": "dist/client-side-validations.js",
   "module": "dist/client-side-validations.esm.js",
@@ -54,14 +48,6 @@
     "dist",
     "src/**/*.js"
   ],
-  "standard": {
-    "ignore": [
-      "/dist/",
-      "/gemfiles/",
-      "/test/",
-      "/vendor/"
-    ]
-  },
   "version": "25.0.0",
   "directories": {
     "lib": "lib",
